+++ ChangeLog.md
@@ -1,5 +1,23 @@
 ## ChangeLog
 
+### v6.5 - 2022-01-28
+
+@smitmartijn:
+
+- \[new] Match PowervRNI version to indicate support for vRNI versions
+- \[new] Support for NSXALB Data Source (NSX Advanced Load Balancer)
+- \[new] New message groups for the databus subscriber: flows, metrics, VMs, hosts, NICs, switchports
+- \[new] Get-vRNISDDCGroup
+- \[new] Get-vRNIVMCDirectConnect
+- \[new] Get-vRNIVMCDirectConnectInterface
+- \[new] Get-vRNISwitchPort
+- \[new] Get-vRNILogicalRouter
+- \[new] Get-vRNIVMwareTransitGateway
+- \[new] Get-vRNINSXTIPsecVPNSessions
+- \[new] Get-vRNISettingsLoginBanner
+- \[new] Set-vRNISettingsLoginBanner
+- \[new] Remove-vRNISettingsLoginBanner
+
 ### v2.0 - 2021-10-05
 
 @smitmartijn: